Made the weights a single array instead of a 2d array for SIMD optimization in calc grad activation func

+++ dataloaders/idx-file-parser.c
@@ -142,7 +142,7 @@ struct pixel_data* get_image_pixel_data(FILE* file) {
     }
     // Read pixel data
     for (uint32_t i = 0; i < size; i++) {
-        size_t bytes_read = fread(neuron_activations->neuron_activation[i], sizeof(uint8_t), rows * cols, file);
+        uint32_t bytes_read = fread(neuron_activations->neuron_activation[i], sizeof(uint8_t), rows * cols, file);
         if (bytes_read != rows * cols) {
             printf("Error: Not enough bytes read for image %u. Expected %u, got %u.\n", i, rows * cols, bytes_read);
             fclose(file);
